common build/watch script
