+++ package.json
@@ -29,15 +29,9 @@
         "clean": "rm -rf frontend/dist && mkdir frontend/dist",
         "build": "yarn clean && yarn build:esbuild && yarn copy-scripts && python manage.py collectstatic --noinput",
         "build:webpack": "echo \"Building Webpack\" && NODE_ENV=production webpack --config webpack.config.js && cp -a frontend/public/* frontend/dist/",
-        "build:esbuild": "yarn build:esbuild:app && yarn build:esbuild:toolbar && yarn build:esbuild:shared_dashboard",
-        "build:esbuild:app": "echo \"🔧 Building App\" && node frontend/build-app.mjs",
-        "build:esbuild:toolbar": "echo \"🔧 Building Toolbar\" && node frontend/build-toolbar.mjs",
-        "build:esbuild:shared_dashboard": "echo \"🔧 Building Shared Dahsboards\" && node frontend/build-shared-dashboard.mjs",
+        "build:esbuild": "node frontend/build.mjs",
         "watch": "yarn watch:esbuild",
-        "watch:esbuild": "concurrently -n app,toolbar,dashboards -c blue,magenta,yellow,red \"yarn watch:esbuild:app\" \"yarn watch:esbuild:toolbar\" \"yarn watch:esbuild:shared_dashboard\"",
-        "watch:esbuild:app": "echo \"👀 Watching App\" && node frontend/build-app.mjs --watch --dev",
-        "watch:esbuild:toolbar": "echo \"👀 Watching Toolbar\" && node frontend/build-toolbar.mjs --watch --dev",
-        "watch:esbuild:shared_dashboard": "echo \"👀 Watching Shared Dahsboards\" && node frontend/build-shared-dashboard.mjs --watch",
+        "watch:esbuild": "node frontend/build.mjs --dev --watch",
         "prettier": "prettier --write \"./**/*.{js,mjs,ts,tsx,json,yaml,yml,css,scss}\"",
         "prettier:check": "prettier --check \"./**/*.{js,mjs,ts,tsx,json,yaml,yml,css,scss}\"",
         "typescript:check": "tsc",
